"arreglos" varios
Estuve parchando un toq lo que rompí en la vez anterior, pero encontré dos problemas nuevos:
 *Al borrar una persona, se borran los registros en Inquilinos y Duenios ahora, pero no verifica que las unidades queden vacías o no.
 *Al registrar un inquilino o duenio nuevo, no reconoce a los que ya existen en la base y los duplica.
 El resto de las cosas debería estar bien.

+++ src/controlador/Controlador.java
@@ -165,7 +165,8 @@ public class Controlador {
 		
 	public void eliminarPersona(String documento) throws PersonaException {
 		Persona persona = buscarPersona(documento);
-		persona.delete();
+		if(persona !=null)
+			persona.delete();
 	}
 	
 	
@@ -240,7 +241,7 @@ public class Controlador {
 		return UnidadDAO.getUnidad(codigo, piso, numero);
 	}	
 	
-	private Persona buscarPersona(String documento) throws PersonaException {
+	public Persona buscarPersona(String documento) throws PersonaException {
 		return PersonaDAO.getPersona(documento);
 	}
